release 0.9.0 commit, changes and features added are in release notes

+++ salt_nornir/runners/nornir_proxy_runner_module.py
@@ -199,7 +199,8 @@ def _run_job(
         events_thread.join(timeout=5)
 
     # kill local client instance
-    client.destroy()
+    if hasattr(client, "destroy"):
+        client.destroy()
 
     return ret
